Turned on "treat warnings as errors" to keep build clean. Tidied up warnings that were now errors. Note in some cases warnings/errors disabled with pragma as need to be kept (for now).

+++ src/ClassicUO.Utility/MathHelper.cs
@@ -61,9 +61,9 @@ namespace ClassicUO.Utility
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static ulong Combine(int val1, int val2)
+        public static ulong Combine(uint val1, uint val2)
         {
-            return (ulong) val1 | ((ulong) val2 << 32);
+            return (val1 | (val2 << 32));
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
